Change timestamps to int64_t

+++ src/xop/media.h
@@ -41,7 +41,7 @@ struct AVFrame
 	std::shared_ptr<uint8_t> buffer; /* 帧数据 */
 	uint32_t size;				     /* 帧大小 */
 	uint8_t  type;				     /* 帧类型 */	
-	uint32_t timestamp;		  	     /* 时间戳 */
+	int64_t timestamp;		  	     /* 时间戳 */
 };
 
 static const int MAX_MEDIA_CHANNEL = 2;
